Just a few suggestions

+++ src/hardware.cpp
@@ -1,5 +1,5 @@
 #include "hardware.h"
 
-float adcToVoltage(short adc) {
+float adcToVoltage(float adc) { // cast to float to avoid rounding errors
     return adc / 1024.0 * 3 / 2;
 }
